Updated document, embeddings, and pinecone service and the visualization references
